Moved table types and hooks to reusable files
